Great document reformatting
- All the comments have been moved to the line above them. This way
it is much easier to see what they refer to on narrow editors, which
almost everyone will have -- one half of monitor for editor, the second
for preview.
- The 'main.tex' has been shuffled. Now, instead of one big 'usepackage'
block and 'settings' block, there are several smaller ones, separated by
their focus (formatting, links, mathematics, etc). This way it is much
easier to find the thing you are looking for.
- Included '\Mod', '\Pmod` and `\Forall` math commands, because their
native versions have spacing flaws.
- Reformatted the 'text.tex' file: tab size is four, the tables are
aligned to make the editing easier.
- I've ensured there are always newlines at the end of the file.